+++ src/main/java/com/ajou_nice/with_pet/admin/service/AdminService.java
@@ -4,12 +4,12 @@ import com.ajou_nice.with_pet.admin.model.admin.AdminApplicantRequest;
 import com.ajou_nice.with_pet.admin.model.admin.AdminApplicantResponse;
 import com.ajou_nice.with_pet.admin.model.criticalservice.AddCriticalServiceRequest;
 import com.ajou_nice.with_pet.admin.model.criticalservice.AddCriticalServiceRequest.CriticalServiceModifyRequest;
-import com.ajou_nice.with_pet.critical_service.dto.CriticalServiceResponse;
+import com.ajou_nice.with_pet.critical_service.model.dto.CriticalServiceResponse;
 import com.ajou_nice.with_pet.admin.model.withpetservice.AddWithPetServiceRequest;
 import com.ajou_nice.with_pet.admin.model.withpetservice.AddWithPetServiceRequest.WithPetServiceModifyRequest;
 import com.ajou_nice.with_pet.admin.model.withpetservice.WithPetServiceResponse;
 import com.ajou_nice.with_pet.admin.util.AdminValidation;
-import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.model.entity.CriticalService;
 import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterBasicResponse;
 import com.ajou_nice.with_pet.domain.dto.petsitterapplicant.ApplicantBasicInfoResponse;
 import com.ajou_nice.with_pet.domain.entity.*;
