+++ src/main/java/com/ajou_nice/with_pet/admin/util/AdminValidation.java
@@ -1,7 +1,7 @@
 package com.ajou_nice.with_pet.admin.util;
 
 import com.ajou_nice.with_pet.admin.model.criticalservice.AddCriticalServiceRequest;
-import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.model.entity.CriticalService;
 import com.ajou_nice.with_pet.domain.entity.User;
 import com.ajou_nice.with_pet.enums.UserRole;
 import com.ajou_nice.with_pet.exception.AppException;
