+++ src/main/java/com/ajou_nice/with_pet/critical_service/repository/CriticalServiceRepository.java
@@ -1,6 +1,6 @@
 package com.ajou_nice.with_pet.critical_service.repository;
 
-import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.model.entity.CriticalService;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
