+++ src/main/java/com/ajou_nice/with_pet/domain/entity/PetSitterCriticalService.java
@@ -9,7 +9,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.model.entity.CriticalService;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
